Items are now ScriptableObjects.
Items were simple calsses, and passing them around was passing copies of their data - wasteful, and this data could be made inconsistent. They were also extremely volatile, since every single item was part of the ItemDatabase on the GameManager prefab. Now that they are scriptable objects, their data is saved as an asset, and they are passed as references.